+++ src/main/java/com/fpt/cpdm/services/UserService.java
@@ -3,6 +3,6 @@ package com.fpt.cpdm.services;
 import com.fpt.cpdm.models.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
-public interface UserService extends UserDetailsService, CRUDService<User, Long> {
+public interface UserService extends UserDetailsService, CRUDService<User, Integer> {
 
 }
